backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.8
python-multipart==0.0.9
Pillow==10.4.0
torch==2.3.1
transformers==4.44.2
sqlalchemy==2.0.32